Move LocalAttribute class and related extension methods to the internal module
