+++ dd-sdk-android-core/api/dd-sdk-android-core.api
@@ -733,38 +733,6 @@ public final class com/datadog/android/core/feature/event/ThreadDump {
 	public fun toString ()Ljava/lang/String;
 }
 
-public abstract interface class com/datadog/android/core/internal/attributes/LocalAttribute {
-}
-
-public abstract interface class com/datadog/android/core/internal/attributes/LocalAttribute$Constant {
-	public abstract fun getKey ()Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-}
-
-public final class com/datadog/android/core/internal/attributes/LocalAttribute$Key : java/lang/Enum {
-	public static final field CREATION_SAMPLING_RATE Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-	public static final field REPORTING_SAMPLING_RATE Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-	public static final field VIEW_SCOPE_INSTRUMENTATION_TYPE Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-	public fun toString ()Ljava/lang/String;
-	public static fun valueOf (Ljava/lang/String;)Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-	public static fun values ()[Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-}
-
-public final class com/datadog/android/core/internal/attributes/LocalAttributeKt {
-	public static final fun enrichWithConstantAttribute (Ljava/util/Map;Lcom/datadog/android/core/internal/attributes/LocalAttribute$Constant;)Ljava/util/Map;
-	public static final fun enrichWithLocalAttribute (Ljava/util/Map;Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;Ljava/lang/Object;)Ljava/util/Map;
-	public static final fun enrichWithNonNullAttribute (Ljava/util/Map;Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;Ljava/lang/Object;)Ljava/util/Map;
-}
-
-public final class com/datadog/android/core/internal/attributes/ViewScopeInstrumentationType : java/lang/Enum, com/datadog/android/core/internal/attributes/LocalAttribute$Constant {
-	public static final field ACTIVITY Lcom/datadog/android/core/internal/attributes/ViewScopeInstrumentationType;
-	public static final field COMPOSE Lcom/datadog/android/core/internal/attributes/ViewScopeInstrumentationType;
-	public static final field FRAGMENT Lcom/datadog/android/core/internal/attributes/ViewScopeInstrumentationType;
-	public static final field MANUAL Lcom/datadog/android/core/internal/attributes/ViewScopeInstrumentationType;
-	public fun getKey ()Lcom/datadog/android/core/internal/attributes/LocalAttribute$Key;
-	public static fun valueOf (Ljava/lang/String;)Lcom/datadog/android/core/internal/attributes/ViewScopeInstrumentationType;
-	public static fun values ()[Lcom/datadog/android/core/internal/attributes/ViewScopeInstrumentationType;
-}
-
 public final class com/datadog/android/core/internal/net/DefaultFirstPartyHostHeaderTypeResolver : com/datadog/android/core/internal/net/FirstPartyHostHeaderTypeResolver {
 	public fun <init> (Ljava/util/Map;)V
 	public fun getAllHeaderTypes ()Ljava/util/Set;
