+++ dd-sdk-android-internal/api/dd-sdk-android-internal.api
@@ -1,3 +1,35 @@
+public abstract interface class com/datadog/android/internal/attributes/LocalAttribute {
+}
+
+public abstract interface class com/datadog/android/internal/attributes/LocalAttribute$Constant {
+	public abstract fun getKey ()Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+}
+
+public final class com/datadog/android/internal/attributes/LocalAttribute$Key : java/lang/Enum {
+	public static final field CREATION_SAMPLING_RATE Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+	public static final field REPORTING_SAMPLING_RATE Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+	public static final field VIEW_SCOPE_INSTRUMENTATION_TYPE Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+	public fun toString ()Ljava/lang/String;
+	public static fun valueOf (Ljava/lang/String;)Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+	public static fun values ()[Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+}
+
+public final class com/datadog/android/internal/attributes/LocalAttributeKt {
+	public static final fun enrichWithConstantAttribute (Ljava/util/Map;Lcom/datadog/android/internal/attributes/LocalAttribute$Constant;)Ljava/util/Map;
+	public static final fun enrichWithLocalAttribute (Ljava/util/Map;Lcom/datadog/android/internal/attributes/LocalAttribute$Key;Ljava/lang/Object;)Ljava/util/Map;
+	public static final fun enrichWithNonNullAttribute (Ljava/util/Map;Lcom/datadog/android/internal/attributes/LocalAttribute$Key;Ljava/lang/Object;)Ljava/util/Map;
+}
+
+public final class com/datadog/android/internal/attributes/ViewScopeInstrumentationType : java/lang/Enum, com/datadog/android/internal/attributes/LocalAttribute$Constant {
+	public static final field ACTIVITY Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
+	public static final field COMPOSE Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
+	public static final field FRAGMENT Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
+	public static final field MANUAL Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
+	public fun getKey ()Lcom/datadog/android/internal/attributes/LocalAttribute$Key;
+	public static fun valueOf (Ljava/lang/String;)Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
+	public static fun values ()[Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
+}
+
 public final class com/datadog/android/internal/collections/EvictingQueue : java/util/Queue {
 	public fun <init> (I)V
 	public synthetic fun <init> (IILkotlin/jvm/internal/DefaultConstructorMarker;)V
