+++ dd-sdk-android-internal/src/main/java/com/datadog/android/internal/attributes/ViewScopeInstrumentationType.kt
@@ -3,14 +3,11 @@
  * This product includes software developed at Datadog (https://www.datadoghq.com/).
  * Copyright 2016-Present Datadog, Inc.
  */
-package com.datadog.android.core.internal.attributes
-
-import com.datadog.android.lint.InternalApi
+package com.datadog.android.internal.attributes
 
 /**
  * A set of constants describing the instrumentation that were used to define the view scope.
  */
-@InternalApi
 enum class ViewScopeInstrumentationType : LocalAttribute.Constant {
     /** Tracked manually through the RUMMonitor API. */
     MANUAL,
