+++ features/dd-sdk-android-rum/src/test/kotlin/com/datadog/android/rum/metric/ViewEndedMetricDispatcherTest.kt
@@ -7,7 +7,7 @@ package com.datadog.android.rum.metric
 
 import com.datadog.android.api.InternalLogger
 import com.datadog.android.api.InternalLogger.Target
-import com.datadog.android.core.internal.attributes.ViewScopeInstrumentationType
+import com.datadog.android.internal.attributes.ViewScopeInstrumentationType
 import com.datadog.android.rum.internal.domain.scope.RumViewType
 import com.datadog.android.rum.internal.metric.NoValueReason
 import com.datadog.android.rum.internal.metric.ViewEndedMetricDispatcher
