+++ integrations/dd-sdk-android-compose/build.gradle.kts
@@ -52,6 +52,7 @@ android {
 }
 
 dependencies {
+    implementation(project(":dd-sdk-android-internal"))
     implementation(project(":features:dd-sdk-android-rum"))
     implementation(libs.kotlin)
